+++ src/components/layout2.tsx
@@ -47,7 +47,12 @@ const Layout2: React.FC<Props> = ({ location, title, children }) => {
         style={{
           marginLeft: `auto`,
           marginRight: `auto`,
-          maxWidth: rhythm(location.pathname === rootPath ? 36 : 24),
+          maxWidth:
+            location.pathname === rootPath
+              ? rhythm(36)
+              : location.pathname.startsWith("/app")
+              ? "var(--geist-page-width-with-margin)"
+              : rhythm(24),
           padding: `${rhythm(1.5)} ${rhythm(3 / 4)}`,
         }}
       >